tests/wpt/tests/content-security-policy/tentative/require-sri-for/script-allowed-meta.https.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<meta http-equiv="Content-Security-Policy" content="require-sri-for 'script'">
<script>
  window.executed_test = async_test("Script that requires integrity executes and does not generate a violation report.");
  document.addEventListener('securitypolicyviolation', executed_test.unreached_func("No report should be generated."));
</script>
<script crossorigin integrity="sha384-tqyFpeo21WFM8HDeUtLqH20GUq/q3D1R6mqTzW3RtyTZ3dAYZJhC1wUcnkgOE2ak"
        src="/content-security-policy/resources/ran.js"></script>
<script>
  assert_true(window.ran);
  window.executed_test.done();
</script>
